Move repository parsing logic to coursier module
