+++ modules/core/shared/src/main/scala/coursier/util/Parse.scala
@@ -1,8 +1,6 @@
 package coursier.util
 
 import coursier.core._
-import coursier.ivy.IvyRepository
-import coursier.maven.MavenRepository
 
 import scala.collection.mutable.ArrayBuffer
 
@@ -301,29 +299,5 @@ object Parse {
                            transitive: Boolean,
                            defaultScalaVersion: String): (Seq[String], Seq[(Dependency, Map[String, String])]) =
     valuesAndErrors(moduleVersionConfig(_, req, transitive, defaultScalaVersion), l)
-
-  def repository(s: String): Either[String, Repository] =
-    if (s == "central")
-      Right(Repositories.central)
-    else if (s.startsWith("sonatype:"))
-      Right(Repositories.sonatype(s.stripPrefix("sonatype:")))
-    else if (s.startsWith("bintray:"))
-      Right(Repositories.bintray(s.stripPrefix("bintray:")))
-    else if (s.startsWith("bintray-ivy:"))
-      Right(Repositories.bintrayIvy(s.stripPrefix("bintray-ivy:")))
-    else if (s.startsWith("typesafe:ivy-"))
-      Right(Repositories.typesafeIvy(s.stripPrefix("typesafe:ivy-")))
-    else if (s.startsWith("typesafe:"))
-      Right(Repositories.typesafe(s.stripPrefix("typesafe:")))
-    else if (s.startsWith("sbt-maven:"))
-      Right(Repositories.sbtMaven(s.stripPrefix("sbt-maven:")))
-    else if (s.startsWith("sbt-plugin:"))
-      Right(Repositories.sbtPlugin(s.stripPrefix("sbt-plugin:")))
-    else if (s.startsWith("ivy:"))
-      IvyRepository.parse(s.stripPrefix("ivy:"))
-    else if (s == "jitpack")
-      Right(Repositories.jitpack)
-    else
-      Right(MavenRepository(s))
 
 }
