+++ modules/coursier/js/src/main/scala/coursier/internal/PlatformRepositoryParser.scala
@@ -1,0 +1,10 @@
+package coursier.internal
+
+import coursier.core.Repository
+
+object PlatformRepositoryParser {
+
+  def repository(input: String): Either[String, Repository] =
+    SharedRepositoryParser.repository(input)
+
+}
