+++ modules/coursier/jvm/src/main/scala/coursier/PlatformResolve.scala
@@ -1,6 +1,7 @@
 package coursier
 
-import coursier.cache.{CacheParse, LocalRepositories}
+import coursier.cache.LocalRepositories
+import coursier.parse.RepositoryParser
 import coursier.util.Repositories
 
 abstract class PlatformResolve {
@@ -14,7 +15,7 @@ abstract class PlatformResolve {
         .toSeq
         .filter(_.nonEmpty)
 
-      CacheParse.repositories(l).either match {
+      RepositoryParser.repositories(l).either match {
         case Left(errs) =>
           System.err.println(
             s"Ignoring $origin, error parsing repositories from it:\n" +
